APD/DI fix, remove uint16 typecast

+++ cardiacmap/transforms.py
@@ -325,15 +325,14 @@ def GetIntersectionsAPD_DI(data, threshold):
     xLen = len(data[0])
 
     # get crossing indices
-
     idx = np.argwhere(np.diff(np.sign(data - threshold))) # this line is by far the most time consuming
     # TODO: Is there a better way?
 
     ys = idx[:, 0]
     xs = idx[:, 1]
+
     validSigs = ys * xLen + xs
     idx0 = idx[:, 2]  # index before
-
     # split the index values by signal
     split_idx = np.argwhere(np.diff(validSigs) != 0).flatten() + 1
     splits0 = np.split(idx0, split_idx)
@@ -342,7 +341,6 @@ def GetIntersectionsAPD_DI(data, threshold):
     # get unique signals from "valid" list
     # must do this AFTER splitting
     validSigs = np.unique(validSigs)
-
     outArr = []
     apdsArr = []
     indices = np.arange(xLen * yLen)
@@ -449,4 +447,4 @@ def NormalizeData(data):
     res += RES_MIN
 
     res = np.moveaxis(res, -1, 0)
-    return res.astype(np.uint16)
+    return res
